split Simulation in .h and .cpp
